refactor(http): delete apiClient,centralize http exports and update imports
