+++ client/src/components/pages/Store/Cart/hooks/usePaymentMethod.jsx
@@ -1,8 +1,7 @@
 "use client";
 import { useState, useEffect, useCallback } from "react";
 
-import { httpClient } from "@/lib/http/httpClient";
-import { parseJsonResponse } from "@/lib/http/parseJsonResponse";
+import { httpClient, parseJsonResponse } from "@/lib/http";
 
 export function usePaymentMethods() {
   const [paymentMethods, setPaymentMethods] = useState([]);
