+++ client/src/components/pages/Store/__tests__/StoreLayout.test.jsx
@@ -20,7 +20,7 @@ jest.mock("lucide-react", () => ({
   Languages: () => <div>Languages Icon</div>,
 }));
 
-jest.mock("../../../../lib/http/httpClient", () => ({
+jest.mock("@/lib/http", () => ({
   httpClient: jest.fn(() => Promise.resolve({})),
 }));
