+++ client/src/components/pages/Store/hooks/__tests__/useProducts.test.jsx
@@ -3,16 +3,15 @@ import { act, useEffect } from "react";
 import { render, screen, waitFor } from "@testing-library/react";
 
 import { useUpload } from "@/components/hooks/useUpload";
-import { httpClient } from "@/lib/http/httpClient";
-import { parseJsonResponse } from "@/lib/http/parseJsonResponse";
+import { httpClient, parseJsonResponse } from "@/lib/http";
 
 import { useProducts } from "../useProducts";
 
-jest.mock("@/lib/http/httpClient", () => ({
+jest.mock("@/lib/http", () => ({
   httpClient: jest.fn(),
 }));
 
-jest.mock("@/lib/http/parseJsonResponse", () => ({
+jest.mock("@/lib/http", () => ({
   parseJsonResponse: jest.fn(),
 }));
