+++ client/src/components/pages/Store/hooks/__tests__/useUsers.test.jsx
@@ -2,16 +2,15 @@ import { act, useEffect } from "react";
 
 import { render, screen, waitFor } from "@testing-library/react";
 
-import { httpClient } from "@/lib/http/httpClient";
-import { parseJsonResponse } from "@/lib/http/parseJsonResponse";
+import { httpClient, parseJsonResponse } from "@/lib/http";
 
 import { useUsers } from "../useUsers";
 
-jest.mock("@/lib/http/httpClient", () => ({
+jest.mock("@/lib/http", () => ({
   httpClient: jest.fn(),
 }));
 
-jest.mock("@/lib/http/parseJsonResponse", () => ({
+jest.mock("@/lib/http", () => ({
   parseJsonResponse: jest.fn(),
 }));
