+++ client/src/components/pages/Store/hooks/useRoles.jsx
@@ -1,8 +1,7 @@
 "use client";
 import { useState, useEffect, useCallback } from "react";
 
-import { httpClient } from "@/lib/http/httpClient";
-import { parseJsonResponse } from "@/lib/http/parseJsonResponse";
+import { httpClient, parseJsonResponse } from "@/lib/http";
 
 export function useRoles() {
   const [roles, setRoles] = useState([]);
