+++ client/src/components/pages/Store/hooks/useTemplates.jsx
@@ -2,8 +2,7 @@
 import { useState, useEffect, useCallback } from "react";
 
 import { toArray } from "@/components/utils/array";
-import { httpClient } from "@/lib/http/httpClient";
-import { parseJsonResponse } from "@/lib/http/parseJsonResponse";
+import { httpClient, parseJsonResponse } from "@/lib/http";
 
 export function useTemplates() {
   const [templates, setTemplates] = useState([]);
