+++ client/src/lib/auth/authSession.js
@@ -1,5 +1,4 @@
-import { httpClient } from "@/lib/http/httpClient";
-import { parseJsonResponse } from "@/lib/http/parseJsonResponse";
+import { httpClient, parseJsonResponse } from "@/lib/http";
 
 export async function getCurrentSession() {
   const response = await httpClient("/users/me");
