+++ client/src/modules/cashier/cashierService.js
@@ -1,5 +1,4 @@
-import { httpClient } from "../../lib/http/httpClient";
-import { parseJsonResponse } from "../../lib/http/parseJsonResponse";
+import { httpClient, parseJsonResponse } from "@/lib/http";
 import { getPaymentByTicketId } from "../orders/ordersService";
 
 export const loginWallet = async (password) => {
