+++ client/src/modules/orders/ordersService.js
@@ -1,4 +1,4 @@
-﻿import { apiClient } from "../../services/apiClient";
+//import { apiClient } from "../../services/apiClient";
 
 export async function getAllOrders() {
   const response = await apiClient("/orders");
